+++ sources/search_pt_segment.c
@@ -6,14 +6,14 @@
 /*   By: cdrouet <cdrouet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/02 08:12:05 by cdrouet           #+#    #+#             */
-/*   Updated: 2017/03/03 15:08:02 by cdrouet          ###   ########.fr       */
+/*   Updated: 2017/03/03 15:19:36 by cdrouet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "wolf3d.h"
 #include <stdio.h>
 
-int				verif_wall(int x, int y, char **map)
+int					verif_wall(int x, int y, char **map)
 {
 	if ((y / BLOCK_SIZE) >= tab_len(map) || y < 0 ||
 			(x / BLOCK_SIZE) >= (int)ft_strlen(map[0]) || x < 0)
@@ -23,11 +23,11 @@ int				verif_wall(int x, int y, char **map)
 
 static t_double_pts	horizontal_search(t_all *all, double angle)
 {
-	t_double_pts		pos;
-	double		xa;
-	double		ya;
-	double	tmpx;
-	double	tmpy;
+	t_double_pts	pos;
+	double			xa;
+	double			ya;
+	double			tmpx;
+	double			tmpy;
 
 	tmpy = ((int)(all->player->pos.y / BLOCK_SIZE) * BLOCK_SIZE) +
 		((angle >= 0 && angle < 180) ? -1 : BLOCK_SIZE);
@@ -35,8 +35,8 @@ static t_double_pts	horizontal_search(t_all *all, double angle)
 		/ tan(RAD(angle));
 	ya = (angle >= 0 && angle < 180) ? -BLOCK_SIZE : BLOCK_SIZE;
 	xa = ((angle >= 180) ? -1 : 1) * (BLOCK_SIZE / tan(RAD(angle)));
-	while (tmpy >= 0 && tmpy < (tab_len(all->map) * BLOCK_SIZE)
-		&& tmpx >= 0 && tmpx < (int)(ft_strlen(all->map[0]) * BLOCK_SIZE)
+	while (tmpy >= 0 && tmpy < (double)(tab_len(all->map) * BLOCK_SIZE)
+		&& tmpx >= 0 && tmpx < (double)(ft_strlen(all->map[0]) * BLOCK_SIZE)
 		&& !verif_wall(tmpx, tmpy, all->map))
 	{
 		tmpx += xa;
@@ -49,11 +49,11 @@ static t_double_pts	horizontal_search(t_all *all, double angle)
 
 static t_double_pts	vertical_search(t_all *all, double angle)
 {
-	t_double_pts		pos;
-	double		xa;
-	double		ya;
-	double	tmpx;
-	double	tmpy;
+	t_double_pts	pos;
+	double			xa;
+	double			ya;
+	double			tmpx;
+	double			tmpy;
 
 	tmpx = ((int)(all->player->pos.x / BLOCK_SIZE) * BLOCK_SIZE) +
 		((angle >= 90 && angle < 270) ? -1 : BLOCK_SIZE);
@@ -62,8 +62,8 @@ static t_double_pts	vertical_search(t_all *all, double angle)
 	xa = (angle >= 90 && angle < 270) ? -BLOCK_SIZE : BLOCK_SIZE;
 	ya = ((angle >= 90 && angle < 270) ? 1 : -1) *
 			(BLOCK_SIZE * tan(RAD(angle)));
-	while (tmpy >= 0 && tmpy < (tab_len(all->map) * BLOCK_SIZE)
-		&& tmpx >= 0 && tmpx < (int)(ft_strlen(all->map[0]) * BLOCK_SIZE)
+	while (tmpy >= 0 && tmpy < (double)(tab_len(all->map) * BLOCK_SIZE)
+		&& tmpx >= 0 && tmpx < (double)(ft_strlen(all->map[0]) * BLOCK_SIZE)
 		&& !verif_wall(tmpx, tmpy, all->map))
 	{
 		tmpx += xa;
@@ -74,7 +74,7 @@ static t_double_pts	vertical_search(t_all *all, double angle)
 	return (pos);
 }
 
-double			search_pts_in_space(t_all *all, double angle)
+double				search_pts_in_space(t_all *all, double angle)
 {
 	t_double_pts	a;
 	t_double_pts	b;
@@ -86,7 +86,7 @@ double			search_pts_in_space(t_all *all, double angle)
 	a = horizontal_search(all, angle);
 	pa = hypot((double)all->player->pos.x - a.x, (double)all->player->pos.y - a.y);
 	b = vertical_search(all, angle);
-	pb = hypot(all->player->pos.x - b.x, all->player->pos.y - b.y);
+	pb = hypot((double)all->player->pos.x - b.x, (double)all->player->pos.y - b.y);
 	if (angle == 180 || angle == 0 || a.x < 0 || a.y < 0
 			|| a.x >= (double)(ft_strlen(all->map[0]) * BLOCK_SIZE)
 			|| a.y >= (double)(tab_len(all->map) * BLOCK_SIZE))
